pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-salesforce-connector"
version = "0.1.2"
description = "A Model Context Protocol (MCP) server implementation for Salesforce integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp",
    "simple-salesforce",
    "python-dotenv",
]
license = { text = "MIT" }
authors = [
    { name = "Suman G", email = "suman.gunaganti@gmail.com" },
]

[project.urls]
Homepage = "https://github.com/smn2gnt/MCP-Salesforce"

[project.scripts]
salesforce = "src.salesforce:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]
